# Remove an expensive call to flatten in `ndarray.init*`