--- a/.travis.yml
+++ b/.travis.yml
@@ -7,7 +7,6 @@ script: bash -ex ./.travis-docker.sh
 
 env:
   global:
-    - EIGENCPP_OPTFLAGS="-Wno-ignored-attributes -O3 -Ofast -march=native -mfpmath=sse -funroll-loops -ffast-math"
     - PACKAGE="owl"
     - PINS="owl-base:. owl-plplot:. owl-top:. owl-zoo:. owl:."