I hope this will mitigate https://github.com/owlbarn/owl/issues/453 (@pveber would you mind checking, I cannot run the benchmark at the moment).

Please be careful when reviewing. What was the reason of going through flatten in the first instance?