@mseri 2019-12-16
I have found a moment to run the benchmarks.

With this patch:

```
❯ dune exec ./owl_perfs.exe mat-mat-mul-small
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌──────────────────────┬──────────┬─────────┬────────────┐
│ Name                 │ Time/Run │ mWd/Run │ Percentage │
├──────────────────────┼──────────┼─────────┼────────────┤
│ lacaml-mat-mat-mul-4 │ 352.58ns │  18.00w │     51.70% │
│ owl-mat-mat-mul-4    │ 681.94ns │ 148.00w │    100.00% │
└──────────────────────┴──────────┴─────────┴────────────┘


❯ dune exec ./owl_perfs.exe vec-init-small
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌───────────────────┬──────────┬─────────┬────────────┐
│ Name              │ Time/Run │ mWd/Run │ Percentage │
├───────────────────┼──────────┼─────────┼────────────┤
│ lacaml-vec-init-4 │ 133.54ns │   9.00w │     43.99% │
│ owl-vec-init-4    │ 303.57ns │  28.00w │    100.00% │
└───────────────────┴──────────┴─────────┴────────────┘
```

Without this patch:
```
❯ dune exec ./owl_perfs.exe vec-init-small
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌───────────────────┬──────────┬─────────┬────────────┐
│ Name              │ Time/Run │ mWd/Run │ Percentage │
├───────────────────┼──────────┼─────────┼────────────┤
│ lacaml-vec-init-4 │ 133.63ns │   9.00w │     38.13% │
│ owl-vec-init-4    │ 350.44ns │  32.00w │    100.00% │
└───────────────────┴──────────┴─────────┴────────────┘

~/code/tests/owl-perfs master 22s
❯ dune exec ./owl_perfs.exe vec-init-large
Estimated testing time 20s (2 benchmarks x 10s). Change using '-quota'.
┌─────────────────────┬────────────┬─────────┬────────────┐
│ Name                │   Time/Run │ mWd/Run │ Percentage │
├─────────────────────┼────────────┼─────────┼────────────┤
│ lacaml-vec-init-100 │   399.91ns │   9.00w │     31.22% │
│ owl-vec-init-100    │ 1_280.81ns │  32.00w │    100.00% │
└─────────────────────┴────────────┴─────────┴────────────┘
```