@mseri 2019-12-16
On https://discuss.ocaml.org/t/working-with-a-huge-data-chunks/3955/10 I have read
```
Creating a view, e.g., via reshape or sub is usually slow, because it is a custom value and has to be created in the major heap, thus every creating of a bigarray triggers the minor collection. We were addressing those issues by preparing all the views that we need in advance (or having a pool of views). [...]
```
Maybe this is really the culprit and we need to find an alternative mechanism.