@pveber 2019-12-20
Thanks for this attempt @mseri! I was wondering: the underlying type for matrices could be a sum type, that is a `Mat.t` could be:
```ocaml
type t =
  | Gen of (...) Genarray.t
  | Mat2 of (...) Array2.t
  | ...
```
When using only lapack operations, we could then avoid all overhead while keeping the API unchanged (and simple).

On a completely orthogonal note, I have found *extremely* useful to [type vectors and matrices operations in a rather strict way](https://github.com/biocaml/phylogenetics/blob/b7ca01042bb53ba09158307063f8c0237faca9f6/lib/linear_algebra.mli), it showed me lots of bugs compared to using `owl` API directly.