@mseri 2019-12-20
> On a completely orthogonal note, I have found _extremely_ useful to [type vectors and matrices operations in a rather strict way](https://github.com/biocaml/phylogenetics/blob/b7ca01042bb53ba09158307063f8c0237faca9f6/lib/linear_algebra.mli), it showed me lots of bugs compared to using `owl` API directly.

I think it is worth opening an issue about this (See now https://github.com/owlbarn/owl/issues/480). I have also been coming back and forth to this issueFrom time to time.