@mseri 2019-12-20
> the underlying type for matrices could be a sum type, that is a `Mat.t` could be:
> 
> ```ocaml
> type t =
>   | Gen of (...) Genarray.t
>   | Mat2 of (...) Array2.t
>   | ...
> ```
> 
> When using only lapack operations, we could then avoid all overhead while keeping the API unchanged (and simple).

This is an interesting idea. It may be worth to try and give it a go at a certain stage. We are in principle free to rebind operations in the modules with more efficient ones. Lots of this code could actually be auto-generated with a suitable ppx