@nilsbecker 2020-01-13
An orthogonal dimension where typing could in principle help improve numerical efficiency is to include matrix structure in the types. LAPACK/BLAS have specialized routines for diagonalizing/decomposing e.g. symmetric or hermitian matrices. some matrix functions are known to preserve these properties, which could also be propagated in their (GAD)type in principle. e.g. eigenvalues of hermitian matrices are real. i think there is/was a ruby numerical library that did something like this, but could not confirm this now when googling Numo.

of course, a somewhat overlapping issue is that matrices with a certain structure can also be stored in various optimal ways (e.g. bands, only one half of symmetric matrices, etc). afaik it's an open question whether a system that tries to automatically dispatch matrices to/from different sparse storage formats according to their structure, can be useful. it might be doing more harm than good by making memory usage unpredictable.